Sketch solution.
