+++ .vscode/launch.json
@@ -16,7 +16,13 @@
 			"outFiles": [
 				"${workspaceFolder}/out/**/*.js"
 			],
-			"preLaunchTask": "npm: watch"
+			"preLaunchTask": "npm: watch",
+			"env": {
+				"PATH": "${env:HOME}/elrondsdk/vendor-rust/bin:${env:HOME}/elrondsdk/erdpy-venv/bin:${env:HOME}/elrondsdk/arwentools:${env:PATH}",
+				"VIRTUAL_ENV": "${env:HOME}/elrondsdk/erdpy-venv",
+				"RUSTUP_HOME": "${env:HOME}/elrondsdk/vendor-rust",
+				"CARGO_HOME": "${env:HOME}/elrondsdk/vendor-rust"
+			}
 		},
 		{
 			"name": "Extension Tests",
