+++ env.vars
@@ -1,0 +1,5 @@
+unset PYTHONHOME
+export PATH=${HOME}/elrondsdk/vendor-rust/bin:${HOME}/elrondsdk/erdpy-venv/bin:${HOME}/elrondsdk/arwentools:${PATH}
+export VIRTUAL_ENV=${HOME}/elrondsdk/erdpy-venv
+export RUSTUP_HOME=${HOME}/elrondsdk/vendor-rust
+export CARGO_HOME=${HOME}/elrondsdk/vendor-rust
